+++ pkg/channels/channels.go
@@ -7,7 +7,7 @@ import (
 	"time"
 )
 
-// OrDone iterates over channel c until it closes or `done` receives a message
+// OrDone forwards data of channel c until c or done close
 func OrDone(done, c <-chan struct{}) <-chan struct{} {
 	valStream := make(chan struct{})
 	go func() {
@@ -30,8 +30,7 @@ func OrDone(done, c <-chan struct{}) <-chan struct{} {
 	return valStream
 }
 
-// OrDoneTimeout iterates over channel c until: (1) c closes, (2) timeout happens, (3) done receives a message
-// closing c means that the initialization procedures has finished
+// OrDoneTimeout forwards data of channel c until: (1) c or done close, (2) timeout happens
 func OrDoneTimeout(done <-chan struct{}, timeout <-chan time.Time, c <-chan struct{}) <-chan struct{} {
 	valStream := make(chan struct{})
 	go func() {
@@ -42,10 +41,13 @@ func OrDoneTimeout(done <-chan struct{}, timeout <-chan time.Time, c <-chan stru
 				return
 			case <-done:
 				return
-			case _, ok := <-c:
+			case v, ok := <-c:
 				if !ok {
-					valStream <- struct{}{}
 					return
+				}
+				select {
+				case valStream <- v:
+				case <-done:
 				}
 			}
 		}
@@ -54,6 +56,7 @@ func OrDoneTimeout(done <-chan struct{}, timeout <-chan time.Time, c <-chan stru
 }
 
 // FanIn merges channels into one output channel
+// Data is forwarded and output channel stays open until all channels close
 func FanIn(done <-chan struct{}, channels ...<-chan struct{}) <-chan struct{} {
 	var wg sync.WaitGroup
 	out := make(chan struct{})
@@ -78,7 +81,10 @@ func FanIn(done <-chan struct{}, channels ...<-chan struct{}) <-chan struct{} {
 	return out
 }
 
-// Or returns when the first of the channels returns
+// Or returns a channel that closes, when any of the input channels close.
+// Data is not forwarded.
+//
+// Example: Merge multiple done channels into a single one
 func Or(channels ...<-chan struct{}) <-chan struct{} {
 	switch len(channels) {
 	case 0:
@@ -106,4 +112,51 @@ func Or(channels ...<-chan struct{}) <-chan struct{} {
 		}
 	}()
 	return orDone
+}
+
+// Barrier returns a channel that closes, when the done channel or any of the input channels close
+// Channel emits only after all input channels have emitted data
+//
+// Example: Merge multiple start-up channels into a single one
+func Barrier(done <-chan struct{}, channels ...<-chan struct{}) <-chan struct{} {
+	var wg sync.WaitGroup
+	var m sync.Mutex
+	inProgress := len(channels)
+	out := make(chan struct{})
+	output := func(c <-chan struct{}) {
+		defer wg.Done()
+		first := true
+		for {
+			select {
+			case <-done:
+				return
+			case v, ok := <-c:
+				if !ok {
+					return
+				}
+				if first {
+					first = false
+					m.Lock()
+					inProgress -= 1
+					// Emit to out channel only after all channels have emitted once
+					if inProgress == 0 {
+						select {
+						case out <- v:
+						case <-done:
+						}
+					}
+					m.Unlock()
+				}
+			}
+		}
+	}
+	wg.Add(len(channels))
+	for _, c := range channels {
+		go output(c)
+	}
+	go func() {
+		wg.Wait()
+		close(out)
+	}()
+	return out
 }