+++ pkg/db/gorm.go
@@ -40,6 +40,7 @@ func Initialize(runCtx context.Context, opts *ConnectionOptions) <-chan struct{}
 	go func() {
 		defer close(dbUp)
 		if opts == nil {
+			dbUp <- struct{}{} // nothing to be done, so show success
 			return
 		}
 		connectFn := func() (*gorm.DB, error) { return connect(opts) }
@@ -51,8 +52,21 @@ func Initialize(runCtx context.Context, opts *ConnectionOptions) <-chan struct{}
 			return
 		}
 		logging.LogInfof("connection to the database succeeded")
+
+		// goroutine to close DB connection when run context is canceled
+		go func() {
+			<-runCtx.Done()
+			logging.LogInfof("run context canceled, closing database connection")
+			defer Close(conn)
+			defer logging.LogInfof("database connection closed")
+		}()
+
 		err = runMigration(conn, opts.MigrationFunc, opts.MigrationVersion)
 		if err != nil {
+			if opts.MigrationHaltOnError {
+				logging.LogErrorf(err, "database migration failed - aborting")
+				return
+			}
 			logging.LogWarningf(err, "database migration failed - continuing")
 		}
 		logging.LogInfof("database migration finished")
@@ -65,15 +79,8 @@ func Initialize(runCtx context.Context, opts *ConnectionOptions) <-chan struct{}
 			registerInstrumenterPlugin()
 			logging.LogInfof("database instrumenter plugin registered")
 		}
-		dbUp <- struct{}{} // notify that DB is up now
-	}()
 
-	// goroutine to close DB connection when run context is canceled
-	go func() {
-		<-runCtx.Done()
-		logging.LogInfof("run context canceled, closing database connection")
-		defer Close(db)
-		defer logging.LogInfof("database connection closed")
+		dbUp <- struct{}{} // notify that DB is up now
 	}()
 	return dbUp
 }